Capstone project UI/Final Project UI/requirements.txt
flask
torch
torchvision
timm
facenet_pytorch
opencv-python
pillow
numpy
matplotlib
